Fix building on v0.30.0
Since Hyprland commit
<https://github.com/hyprwm/Hyprland/commit/3785defaf12b9d99137b2f4c74ab82c51cf733e1>
building the plugin was failing with `CHyprGroupBarDecoration` not being
declared. This happened because this commit has removed an include in
`layout/DwindleLayout.hpp` that contained this struct.

Instead of relying on that include being in `DwindleLayout.hpp`, we now
include it ourselves directly from
`render/decorations/CHyprGroupBarDecoration.hpp`.

Note that even with this fix, the plugin will not be buildable until
Hyprland commit
<https://github.com/hyprwm/Hyprland/commit/1925e64c21811ce76e5059d7a063f968c2d3e98c>
which fixes another issue created by the commit above, where the
`helpers/Vector2D.hpp` had a typo in `macros.hpp` include, which
should've been `../macros.hpp`.

That said, this commit is a part of version 0.30.0, where the plugin is
now working without issues.

Fixes #7